backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
slowapi==0.1.9
redis==5.0.1
structlog==23.2.0
python-json-logger==2.0.7
email-validator==2.1.0
httpx==0.25.2

